+++ Sources/Utilities/SystemMessage.swift
@@ -108,6 +108,7 @@ struct SystemMessage: Localizable {
 
     struct UIError {
         static let unspecifiedLocation = localizedString(forKey: "UnspecifiedLocation")
+        static let InvalidDatErrorInForm = localizedString(forKey: "InvalidDatErrorInForm")
     }
 
     struct PhotoAlbum {
@@ -162,6 +163,9 @@ struct SystemMessage: Localizable {
         static let Report = localizedString(forKey: "Report")
         static let You = localizedString(forKey: "You")
         static let Admin = localizedString(forKey: "Admin")
+        static let DatePickerTimeTitle = localizedString(forKey: "DatePickerTimeTitle")
+        static let DatePickerDateTitle = localizedString(forKey: "DatePickerDateTitle")
+        static let DatePickerDateAndTimeTitle = localizedString(forKey: "DatePickerDateAndTimeTitle")
     }
 
     struct Mute {