+++ Sources/Views/ALKFormCell.swift
@@ -8,7 +8,15 @@
 import UIKit
 
 class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate {
+    enum FormData {
+        static let valid = 1
+        static let inValid = 2
+    }
+
     public var tapped: ((_ index: Int, _ name: String, _ formDataSubmit: FormDataSubmit?) -> Void)?
+
+    public var onTapOfDateSelect: ((_ index: Int, _ delegate: ALKDatePickerButtonClickProtocol?, _ datePickerMode: UIDatePicker.Mode, _ identifier: String) -> Void)?
+
     let itemListView = NestedCellTableView()
     var submitButton: CurvedImageButton?
     var identifier: String?
@@ -30,7 +38,7 @@ class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate {
         }
         set(newFormData) {
             guard let key = identifier,
-                let formData = newFormData else { return }
+                  let formData = newFormData else { return }
             formDataCacheStore.set(formData, for: key)
         }
     }
@@ -55,6 +63,33 @@ class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate {
         template = viewModel.formTemplate()
     }
 
+    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
+        guard let key = identifier else {
+            return false
+        }
+        let item = items[textField.tag]
+        var datePickerMode: UIDatePicker.Mode?
+        switch item.type {
+        case .time:
+            datePickerMode = .time
+        case .date:
+            datePickerMode = .date
+        case .dateTimeLocal:
+            datePickerMode = .dateAndTime
+        default:
+            return true
+        }
+
+        guard let pickerMode = datePickerMode,
+              let dateSelectTap = onTapOfDateSelect
+        else {
+            return true
+        }
+
+        dateSelectTap(textField.tag, self, pickerMode, key)
+        return false
+    }
+
     func textFieldDidBeginEditing(_ textField: UITextField) {
         activeTextField = textField
     }
@@ -62,8 +97,8 @@ class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate {
     func textFieldDidEndEditing(_ textField: UITextField) {
         activeTextField = nil
         guard let text = textField.text,
-            !text.trim().isEmpty,
-            let formSubmitData = formData
+              !text.trim().isEmpty,
+              let formSubmitData = formData
         else {
             if let data = formData {
                 data.textFields.removeValue(forKey: textField.tag)
@@ -92,6 +127,9 @@ class ALKFormCell: ALKChatBaseCell<ALKMessageViewModel>, UITextFieldDelegate {
         itemListView.register(ALKFormPasswordItemCell.self)
         itemListView.register(ALKFormSingleSelectItemCell.self)
         itemListView.register(ALKFormMultiSelectItemCell.self)
+        itemListView.register(ALKFormDateItemCell.self)
+        itemListView.register(ALKFormTimeItemCell.self)
+        itemListView.register(ALKFormDateTimeItemCell.self)
     }
 
     private func setUpSubmitButton(title: String) {
@@ -119,12 +157,33 @@ extension ALKFormCell: UITableViewDataSource, UITableViewDelegate {
             cell.item = item
             cell.valueTextField.delegate = self
             cell.valueTextField.tag = indexPath.section
+            if let formDataSubmit = formData,
+               let text = formDataSubmit.textFields[indexPath.section]
+            {
+                cell.valueTextField.text = text
+            } else {
+                cell.valueTextField.text = ""
+            }
+            if let validationField = formData?.validationFields[indexPath.section], validationField == FormData.inValid {
+                let formViewModelTextItem = item as? FormViewModelTextItem
+                cell.errorLabel.text = formViewModelTextItem?.validation?.errorText ?? localizedString(forKey: "InvalidDatErrorInForm", withDefaultValue: SystemMessage.UIError.InvalidDatErrorInForm, fileName: localizedStringFileName)
+                cell.errorLabel.isHidden = false
+            } else {
+                cell.errorLabel.isHidden = true
+            }
             return cell
         case .password:
             let cell: ALKFormPasswordItemCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
             cell.item = item
             cell.valueTextField.delegate = self
             cell.valueTextField.tag = indexPath.section
+            if let formDataSubmit = formData,
+               let text = formDataSubmit.textFields[indexPath.section]
+            {
+                cell.valueTextField.text = text
+            } else {
+                cell.valueTextField.text = ""
+            }
             return cell
         case .singleselect:
             guard let singleselectItem = item as? FormViewModelSingleselectItem else {
@@ -145,8 +204,8 @@ extension ALKFormCell: UITableViewDataSource, UITableViewDelegate {
             cell.item = singleselectItem.options[indexPath.row]
 
             if let formDataSubmit = formData,
-                let singleSelectFields = formDataSubmit.singleSelectFields[indexPath.section],
-                singleSelectFields == indexPath.row
+               let singleSelectFields = formDataSubmit.singleSelectFields[indexPath.section],
+               singleSelectFields == indexPath.row
             {
                 cell.accessoryType = .checkmark
             } else {
@@ -181,13 +240,62 @@ extension ALKFormCell: UITableViewDataSource, UITableViewDelegate {
             }
 
             if let formDataSubmit = formData,
-                let multiSelectFields = formDataSubmit.multiSelectFields[indexPath.section], multiSelectFields.contains(indexPath.row)
+               let multiSelectFields = formDataSubmit.multiSelectFields[indexPath.section], multiSelectFields.contains(indexPath.row)
             {
                 cell.accessoryType = .checkmark
             } else {
                 cell.accessoryType = .none
             }
             cell.item = multiselectItem.options[indexPath.row]
+            return cell
+
+        case .date:
+            guard let dateSelectItem = item as? FormViewModelDateItem else {
+                return UITableViewCell()
+            }
+            let cell: ALKFormDateItemCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+            cell.item = dateSelectItem
+            cell.valueTextField.delegate = self
+            cell.valueTextField.tag = indexPath.section
+            if let timeInMillSecs = formData?.dateFields[indexPath.section] {
+                let dateFormate = Date.is24HrsFormate() ? Date.Formates.Date.twentyfour : Date.Formates.Date.twelve
+                cell.valueTextField.text = Date.formatedDate(formateString: dateFormate, timeInMillSecs: timeInMillSecs)
+            } else {
+                cell.valueTextField.text = ""
+            }
+            return cell
+        case .time:
+            guard let timeSelectItem = item as? FormViewModelTimeItem else {
+                return UITableViewCell()
+            }
+            let cell: ALKFormTimeItemCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+            cell.item = timeSelectItem
+            cell.valueTextField.delegate = self
+            cell.valueTextField.tag = indexPath.section
+
+            if let timeInMillSecs = formData?.dateFields[indexPath.section] {
+                let timeFormate = Date.is24HrsFormate() ? Date.Formates.Time.twentyfour : Date.Formates.Time.twelve
+                cell.valueTextField.text = Date.formatedDate(formateString: timeFormate, timeInMillSecs: timeInMillSecs)
+            } else {
+                cell.valueTextField.text = ""
+            }
+            return cell
+        case .dateTimeLocal:
+            guard let dateTimeSelectItem = item as? FormViewModelDateTimeLocalItem else {
+                return UITableViewCell()
+            }
+
+            let cell: ALKFormDateTimeItemCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+            cell.item = dateTimeSelectItem
+            cell.valueTextField.delegate = self
+            cell.valueTextField.tag = indexPath.section
+
+            if let timeInMillSecs = formData?.dateFields[indexPath.section] {
+                let dateTimeFromate = Date.is24HrsFormate() ? Date.Formates.DateAndTime.twentyfour : Date.Formates.DateAndTime.twelve
+                cell.valueTextField.text = Date.formatedDate(formateString: dateTimeFromate, timeInMillSecs: timeInMillSecs)
+            } else {
+                cell.valueTextField.text = ""
+            }
             return cell
         }
     }
@@ -216,6 +324,76 @@ extension ALKFormCell: Tappable {
     }
 }
 
+extension ALKFormCell: ALKDatePickerButtonClickProtocol {
+    func confirmButtonClick(position: Int,
+                            date: Date,
+                            messageKey: String,
+                            datePickerMode: UIDatePicker.Mode)
+    {
+        guard identifier == messageKey else { return }
+
+        var timeInMillSecs: Int64 = 0
+
+        switch datePickerMode {
+        case .time:
+            timeInMillSecs = Int64(date.timeIntervalSince1970 * 1000)
+        case .date:
+            timeInMillSecs = Int64(date.timeIntervalSince1970 * 1000)
+        case .dateAndTime:
+            timeInMillSecs = Int64(date.timeIntervalSince1970 * 1000)
+        default:
+            break
+        }
+        guard let formSubmitData = formData,
+              timeInMillSecs > 0,
+              position < itemListView.numberOfSections
+        else {
+            print("Can't be updated due to incorrect index")
+            return
+        }
+        formSubmitData.dateFields[position] = timeInMillSecs
+        formData = formSubmitData
+        itemListView.reloadSections([position], with: .fade)
+    }
+}
+
+extension ALKFormCell {
+    func isFormDataValid() -> Bool {
+        var isValid: Bool = true
+
+        guard let formDataSubmit = formData,
+              let viewModelItems = template?.viewModeItems
+        else {
+            return false
+        }
+        // Loop and match all the text types for validation and mark them as valid or inValid.
+        for index in 0 ..< viewModelItems.count {
+            let element = viewModelItems[index]
+
+            switch element.type {
+            case .text:
+                let textFieldModel = element as? FormViewModelTextItem
+                let enteredText = formDataSubmit.textFields[index] ?? ""
+
+                if let validation = textFieldModel?.validation,
+                   let regxPattern = validation.regex
+                {
+                    do {
+                        isValid = try ALKRegexValidator.matchPattern(text: enteredText, pattern: regxPattern)
+                        formDataSubmit.validationFields[index] = isValid ? FormData.valid : FormData.inValid
+                        formData = formDataSubmit
+                    } catch {
+                        print("Error while matching text: \(error.localizedDescription)")
+                    }
+                }
+            default:
+                break
+            }
+        }
+        return isValid
+    }
+}
+
 class NestedCellTableView: UITableView {
     override var intrinsicContentSize: CGSize {
         self.layoutIfNeeded()
@@ -233,4 +411,6 @@ class FormDataSubmit {
     var textFields = [Int: String]()
     var singleSelectFields = [Int: Int]()
     var multiSelectFields = [Int: [Int]]()
+    var dateFields = [Int: Int64]()
+    var validationFields = [Int: Int]()
 }